Added support for Auto-Upload on Save for individual deployments
This commit includes a `BufWritePost` event handler which first checks
if a deployment exists for the file being saved, then checks if
`upload_on_save` is truthy for the deployment. This is preferable to a
setting in the pligin config because in practice this requirement can be
unique to each mapping within a project.

The `BufWritePost` event handler includes a debouncing approach to
prevent concurrent uploads of the same file due to the non-blocking
nature of the `upload_file` plugin method.